tester/requirements.txt
# https://docs.openstack.org/python-swiftclient/latest/swiftclient.html#module-swiftclient.client
python-swiftclient
elasticsearch
keystoneauth1
requests
